Restructure token module types + add reject reason
